% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cgeneric_generic0.R
\name{cgeneric_generic0}
\alias{cgeneric_generic0}
\alias{cgeneric_iid}
\title{Build an \code{cgeneric} object for a \code{generic0} model.
See details.}
\usage{
cgeneric_generic0(
  R,
  param,
  constr = TRUE,
  scale = TRUE,
  debug = FALSE,
  useINLAprecomp = TRUE,
  libpath = NULL
)

cgeneric_iid(
  n,
  param,
  constr = FALSE,
  scale = TRUE,
  debug = FALSE,
  useINLAprecomp = TRUE,
  libpath = NULL
)
}
\arguments{
\item{R}{the structure matrix for the model definition.}

\item{param}{length two vector with the parameters
\code{a} and \code{p} for the PC-prior distribution defined from
\deqn{P(\sigma > a) = p}
where \eqn{\sigma} can be interpreted as marginal standard
deviation of the process if scale = TRUE. See details.}

\item{constr}{logical indicating if it is to add a
sum-to-zero constraint. Default is TRUE.}

\item{scale}{logical indicating if it is to scale
the model. See detais.}

\item{debug}{integer, default is zero, indicating the verbose level.
Will be used as logical by INLA.}

\item{useINLAprecomp}{logical, default is TRUE, indicating if it is to
be used the shared object pre-compiled by INLA.
This is not considered if 'libpath' is provided.}

\item{libpath}{string, default is NULL, with the path to the shared object.}

\item{n}{integer required to specify the model size}
}
\value{
a \code{cgeneric} object, see \code{\link[=cgeneric]{cgeneric()}}.
}
\description{
Build data needed to implement a model whose
precision has a conditional precision parameter.
This uses the C interface in \code{\link[INLA:cgeneric]{INLA::cgeneric()}},
that can be used as a linear predictor
model component with \code{\link[INLA:f]{INLA::f()}}.
}
\details{
The precision matrix is defined as
\deqn{Q = \tau R}
where the structure matrix R is supplied by the user
and \eqn{\tau} is the precision parameter.
Following Sørbie and Rue (2014), if scale = TRUE
the model is scaled so that
\deqn{Q = \tau s R}
where \eqn{s} is the geometric mean of the diagonal
elements of the generalized inverse of \eqn{R}.
\deqn{s = \exp{\sum_i \log((R^{-})_{ii})/n}}
If the model is scaled, the geometric mean of the
marginal variances, the diagonal of \eqn{Q^{-1}}, is one.
Therefore, when the model is scaled,
\eqn{\tau} is the marginal precision,
otherwise \eqn{\tau} is the conditional precision.
}
\section{Functions}{
\itemize{
\item \code{cgeneric_iid()}: The \code{\link[=cgeneric_iid]{cgeneric_iid()}} uses the \link{cgeneric_generic0}
with the structure matrix as the identity.

}}
\examples{
## two ways to define an 'iid' model
prior.par <- c(1, 0.5)
ma <- cgeneric("generic0", R = Diagonal(5),
   param = prior.par, constr = FALSE)
mb <- cgeneric("iid", n = 5, param = prior.par)
all.equal(ma, mb)

## Retrieve the precision
prec(ma, theta = log(3))

lamb <- -log(prior.par[2])/prior.par[1]
pmode <- log(2)/lamb
## prior for log(precision) and sigma
par(mfrow = c(1, 2))
plot(function(x)
 exp(prior(ma, theta = matrix(x, nrow=1))),
  -3, 3, n = 601, xlab = "precision", ylab = "density")
rug(pmode, lwd = 3, col = 2)
plot(function(x)
exp(prior(ma,
  theta = matrix(-2*log(x), nrow = 1))+log(2)-log(x)),
  1/100, 10, n = 1000,
  xlab = expression(sigma), ylab = "density")
plot(function(x) dexp(x, lamb), 1/100, 10, n = 1000,
   add = TRUE, lty = 2, col = 2)
rug(exp(-pmode/2), lwd = 3, col = 2)

## structured precision matrix model definition
R <- Matrix(toeplitz(c(2,-1,0,0,0)))
mR <- cgeneric("generic0", R = R,
  scale = FALSE, param = c(1, 0.05))
graph(mR)
prec(mR, theta = 0)
}
\references{
Sigrunn Holbek Sørbye and Håvard Rue (2014).
Scaling intrinsic Gaussian Markov random field priors in
spatial modelling. Spatial Statistics, vol. 8, p. 39-51.
}
