% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cgeneric_utils.R
\name{cgeneric_get}
\alias{cgeneric_get}
\alias{initial.cgeneric}
\alias{mu.cgeneric}
\alias{prior.cgeneric}
\alias{graph.cgeneric}
\alias{prec.cgeneric}
\title{\code{cgeneric_get} is an internal function used by
\code{graph}, \code{pred}, \code{initial}, \code{mu} or \code{prior}
methods for \code{cgeneric}.}
\usage{
cgeneric_get(
  model,
  cmd = c("graph", "Q", "initial", "mu", "log_prior"),
  theta,
  optimize = TRUE
)

\method{initial}{cgeneric}(model)

\method{mu}{cgeneric}(model, theta)

\method{prior}{cgeneric}(model, theta)

\method{graph}{cgeneric}(model, optimize)

\method{prec}{cgeneric}(model, theta, optimize)
}
\arguments{
\item{model}{a \code{cgeneric} object.}

\item{cmd}{an string to specify which model element to get}

\item{theta}{numeric vector with the model parameters.
If missing, the \code{\link[=initial]{initial()}} will be used.}

\item{optimize}{logical indicating if it is to be
returned only the elements and not as a sparse matrix.}
}
\value{
depends on \code{cmd}

numeric scalar (if numeric vector is provided
for theta) or vector (if numeric matrix is provided
for theta).
}
\description{
The \code{generic_get} retrieve a model property specified by
\code{cmd} on an \code{cgeneric} object.
The functions listed below are for each \code{cmd} case.
}
\section{Functions}{
\itemize{
\item \code{initial(cgeneric)}: Retrive the initial parameter(s) of an \code{cgeneric} model.

\item \code{mu(cgeneric)}: Evaluate the mean for an \code{cgeneric} model.

\item \code{prior(cgeneric)}: Evaluate the prior for an \code{cgeneric} model

\item \code{graph(cgeneric)}: Retrieve the graph of an \code{cgeneric} object

\item \code{prec(cgeneric)}: Retrieve the precision of an \code{cgeneric} object

}}
\seealso{
check the examples in \code{\link[=cgeneric_generic0]{cgeneric_generic0()}}
}
