% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inla_utils.R
\name{data.expand}
\alias{data.expand}
\title{function to make Y: for multiple likelihood with INLA}
\usage{
data.expand(data, by.var = FALSE)
}
\arguments{
\item{data}{a \code{data.frame} or matrix.}

\item{by.var}{logical to indicate if the order is by variable}
}
\description{
function to make Y: for multiple likelihood with INLA
}
\details{
extract the joint prior precision for the
latent field at the mode of the hyperparameters
}
\examples{
dataf <- data.frame(y1 = 1:3, y2 = rnorm(3))
data.expand(dataf, by.var = FALSE)
data.expand(dataf, by.var = TRUE)
}
