% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cgeneric_kronecker.R
\name{extraconstr}
\alias{extraconstr}
\alias{kronecker_extraconstr}
\title{Kronecker (product) between \code{extraconstr},
implemented for \code{\link[=kronecker]{kronecker()}} methods.}
\usage{
kronecker_extraconstr(c1, c2, n1, n2)
}
\arguments{
\item{c1, c2}{named list with two elements:
\code{A} and \code{e}, where \code{nrow(A)} should be equal
to \code{length(e)}. This is an extraconstr.}

\item{n1, n2}{integer with each model's length}
}
\value{
extraconstr
}
\description{
Kronecker (product) between \code{extraconstr},
implemented for \code{\link[=kronecker]{kronecker()}} methods.
}
