% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fast_aggregation_check.R
\docType{methods}
\name{fast_aggregation_check}
\alias{fast_aggregation_check}
\title{Fast aggregation check
Compare the observed and modelled aggregated values.}
\usage{
fast_aggregation_check(object, grouping_vars, fun = sum, nsim = 1000)
}
\arguments{
\item{object}{the INLA model}

\item{grouping_vars}{character vector of variable names to group by.}

\item{fun}{function to apply to the aggregated values.}

\item{nsim}{the number of simulation}
}
\description{
Fast aggregation check
Compare the observed and modelled aggregated values.
}
\seealso{
Other checks: 
\code{\link{dispersion_check}()},
\code{\link{distribution_check}()},
\code{\link{fast_aggregation_check,inla-method}},
\code{\link{get_anomaly}()}
}
\concept{checks}
