% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fast_aggregation_check.R,
%   R/fast_distribution_check.R
\docType{methods}
\name{fast_aggregation_check,inla-method}
\alias{fast_aggregation_check,inla-method}
\alias{fast_distribution_check}
\alias{fast_distribution_check,inla-method}
\alias{fast_distribution_check,list-method}
\title{Use simulations to compare the observed distribution with the modelled
distribution}
\usage{
\S4method{fast_aggregation_check}{inla}(object, grouping_vars, fun = sum, nsim = 1000)

fast_distribution_check(object, nsim = 1000)

\S4method{fast_distribution_check}{inla}(object, nsim = 1000)

\S4method{fast_distribution_check}{list}(object, nsim = 1000)
}
\arguments{
\item{object}{the INLA model}

\item{nsim}{the number of simulation}
}
\description{
This check uses the fitted values and thus ignores the uncertainty on the
predictions
}
\examples{
library(INLA)
set.seed(20181202)
model <- inla(
  poisson ~ 1,
  family = "poisson",
  data = data.frame(
    poisson = rpois(20, lambda = 10),
    base = 1
  ),
  control.predictor = list(compute = TRUE)
)
fast_distribution_check(model)
}
\seealso{
Other checks: 
\code{\link{dispersion_check}()},
\code{\link{distribution_check}()},
\code{\link{fast_aggregation_check}()},
\code{\link{get_anomaly}()}
}
\concept{checks}
