% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{findAndGetFunction}
\alias{findAndGetFunction}
\title{Search a function and retrieve it.}
\usage{
findAndGetFunction(fName, package, debug = FALSE)
}
\arguments{
\item{fName}{character with the name of the function}

\item{package}{character with the package name}

\item{debug}{logical indicating if it is to print
intermediate progress finding}
}
\value{
function
}
\description{
Search a function and retrieve it.
}
\details{
The (first) package name where it was found
is returned as an atribute named "package"
}
