% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{findGetFunction}
\alias{findGetFunction}
\title{Search a function and retrieve it.}
\usage{
findGetFunction(fName, package, debug = FALSE)
}
\arguments{
\item{fName}{character with the name of the function}

\item{package}{character with the package name}

\item{debug}{logical indicating if it is to print
intermediate progress finding}
}
\value{
function.
The (first) package name where it was found
is returned as an attribute named "package"
}
\description{
Search a function and retrieve it.
}
\details{
if 'missing(package)' it will search on the loaded
packages, first in the exported functions, and then
among the non-exported ones.
NOTE: 'package' can include any installed package,
see \code{\link[=installed.packages]{installed.packages()}}
}
