% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_observed.R
\docType{methods}
\name{fitted,inla-method}
\alias{fitted,inla-method}
\title{Calculate the residuals from an INLA model}
\usage{
\S4method{fitted}{inla}(object, ...)
}
\arguments{
\item{object}{an object for which the extraction of model fitted values is
    meaningful.}

\item{...}{other arguments.}
}
\description{
Calculate the residuals from an INLA model
}
