% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_anomaly.R
\docType{methods}
\name{get_anomaly}
\alias{get_anomaly}
\alias{get_anomaly,inla-method}
\title{Get a set of anomalies of the model}
\usage{
get_anomaly(object, n = 10)

\S4method{get_anomaly}{inla}(object, n = 20)
}
\arguments{
\item{object}{the INLA model}

\item{n}{the number of anomalies per criterion.
Defaults to \code{10}.}
}
\description{
Returns a named list with an element called \code{observations} and one element
for every random effect.
The random effect components use the name of the random effect.
}
\details{
\code{observations} is a subset of the original data.frame.
It contains the rows with the \code{n} largest and \code{n} smallest values of the
Pearson residuals.
The random effect components contain a subset of the random effects.
Here we select the rows with the \code{n} largest and \code{n} lowest values of the
mean.
}
\examples{
library(INLA)
set.seed(20181202)
model <- inla(
  poisson ~ 1,
  family = "poisson",
  data = data.frame(
    poisson = rpois(20, lambda = 10),
    base = 1
  ),
  control.predictor = list(compute = TRUE)
)
dc <- get_anomaly(model, n = 2)
str(dc)
}
\seealso{
Other checks: 
\code{\link{dispersion_check}()},
\code{\link{distribution_check}()},
\code{\link{fast_aggregation_check}()},
\code{\link{fast_aggregation_check,inla-method}}
}
\concept{checks}
