% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_observed.R
\name{get_observed}
\alias{get_observed}
\title{get the observed values from the model object}
\usage{
get_observed(object)
}
\arguments{
\item{object}{the INLA model}
}
\description{
get the observed values from the model object
}
\examples{
library(INLA)
set.seed(20181202)
model <- inla(
  poisson ~ 1,
  family = "poisson",
  data = data.frame(
    poisson = rpois(20, lambda = 10),
    base = 1
  ),
  control.predictor = list(compute = TRUE)
)
get_observed(model)
}
\seealso{
Other statistics: 
\code{\link{dispersion}()},
\code{\link{fitted,inla-method}},
\code{\link{residuals,inla-method}}
}
\concept{statistics}
