% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fast_distribution_check.R
\name{dgpoisson}
\alias{dgpoisson}
\alias{rgpoisson}
\title{The generalised Poisson distribution}
\usage{
dgpoisson(y, mu, phi)

rgpoisson(n, mu, phi)
}
\arguments{
\item{y}{a vector of positive integers for which to calculate the density}

\item{mu}{a vector of averages for which to calculate the density}

\item{phi}{a single overdispersion parameter}

\item{n}{the number of simulated values}
}
\value{
a matrix with the density for each combination of \code{y} (rows) and \code{mu}
(cols)
}
\description{
The generalised Poisson distribution
}
\seealso{
Other statistics: 
\code{\link{dispersion}()},
\code{\link{fitted,inla-method}},
\code{\link{get_observed}()},
\code{\link{residuals,inla-method}}

Other statistics: 
\code{\link{dispersion}()},
\code{\link{fitted,inla-method}},
\code{\link{get_observed}()},
\code{\link{residuals,inla-method}}
}
\concept{statistics}
