% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inla_methods.R
\name{inlaQ}
\alias{inlaQ}
\title{Define the method to extract the precision from an inla output object.}
\usage{
inlaQ(model, ...)
}
\arguments{
\item{model}{an inla output}

\item{...}{used to pass the 'prior' argument,
as logical (default is TRUE) to indicate if it is
to retrieve the prior or the posterior precision,
'id.config' to inform which config it is to be taken,
if not provided it will take the one with highest
log-posterior.}
}
\description{
Define the method to extract the precision from an inla output object.
}
\details{
extract the joint prior precision for the
latent field at the mode of the hyperparameters
}
