% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa_methods.R
\name{INLAtools-methods}
\alias{INLAtools-methods}
\alias{vcov.Matrix}
\title{Methods to work with a \code{model}.}
\usage{
\method{vcov}{Matrix}(object, ...)
}
\arguments{
\item{object}{Matrix supposed to be a
sparse precision matrix}

\item{model}{object to represent a model}
}
\value{
the result of the desired query
of the 'cgeneric' model.
'graph' and 'prec' can be either a vector
(if optimize = TRUE) or a sparse matrix.
}
\description{
For a given model object query the \code{initial},
\code{mu}, log \code{prior}, \code{graph} or precision \code{prec}
can be evaluated/retrieved.
}
\section{Functions}{
\itemize{
\item \code{vcov(Matrix)}: The \code{vcov} method for sparse matrices

}}
