% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa_methods.R
\name{is.zero}
\alias{is.zero}
\alias{is.zero.default}
\alias{is.zero.matrix}
\alias{is.zero.Matrix}
\title{Define the is.zero method}
\usage{
is.zero(x, ...)

\method{is.zero}{default}(x, ...)

\method{is.zero}{matrix}(x, ...)

\method{is.zero}{Matrix}(x, ...)
}
\arguments{
\item{x}{an R object}

\item{...}{additional arguments}
}
\value{
logical
}
\description{
Define the is.zero method
}
\section{Methods (by class)}{
\itemize{
\item \code{is.zero(default)}: The is.zero.default definition

\item \code{is.zero(matrix)}: The is.zero.matrix definition

\item \code{is.zero(Matrix)}: The is.zero.Matrix definition

}}
