% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{packageCheck}
\alias{packageCheck}
\title{To check package version and load}
\usage{
packageCheck(name, minimum_version, quietly = FALSE)
}
\arguments{
\item{name}{character with the name of the package}

\item{minimum_version}{character with the minimum required version}

\item{quietly}{logical indicating if messages shall be printed}
}
\description{
To check package version and load
}
\note{
Original in inlabru package function check_package_version_and_load
}
