% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pkgCheck}
\alias{pkgCheck}
\title{To check package version and load}
\usage{
pkgCheck(pkg, minimum_version, quietly = FALSE)
}
\arguments{
\item{pkg}{character with the name of the package}

\item{minimum_version}{character with the minimum required version}

\item{quietly}{logical indicating if messages shall be printed}
}
\description{
To check package version and load
}
\note{
The original code is in check_package_version_and_load()
function of the 'inlabru' package
}
