% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_distribution_check.R
\name{plot.dispersion_check}
\alias{plot.dispersion_check}
\title{Plot the results from a dispersion check}
\usage{
\method{plot}{dispersion_check}(x, y, ...)
}
\arguments{
\item{x}{a dispersion_check object. Which is the output of
\code{\link{dispersion_check}}}

\item{y}{currently ignored}

\item{...}{currently ignored}
}
\value{
a ggplot2 object
}
\description{
Plot the results from a dispersion check
}
