% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_distribution_check.R
\name{plot.distribution_check}
\alias{plot.distribution_check}
\title{Plot the results from a distribution check}
\usage{
\method{plot}{distribution_check}(x, y, ...)
}
\arguments{
\item{x}{a distribution_check object. Which is the output of
\code{\link{fast_distribution_check}}}

\item{y}{currently ignored}

\item{...}{currently ignored}
}
\value{
a ggplot2 object
}
\description{
Plot the results from a distribution check
}
\seealso{
Other utils: \code{\link{generate_data}},
  \code{\link{plot.dispersion_check}}
}
\concept{utils}
