% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_rw.R
\name{plot.rw_sim}
\alias{plot.rw_sim}
\title{Plot simulated second order random walks}
\usage{
\method{plot}{rw_sim}(x, y, ..., type = c("all", "divergence",
  "stationary", "quantile"), link = c("identity", "log", "logit"))
}
\arguments{
\item{x}{a \code{rw_sim} object. Which is the output of  \code{\link{simulate_rw}}}

\item{y}{currently ignored}

\item{...}{currently ignored}

\item{type}{which plot to create. \code{"all"} displays all simulations. \code{"divergence"} displays the most divergent simulations.}

\item{link}{which link to use for back transformation}
}
\value{
a ggplot2 object
}
\description{
Plot simulated second order random walks
}
\seealso{
Other priors: \code{\link{simulate_rw}}
}
\concept{priors}
