% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_idd.R
\name{plot.sim_iid}
\alias{plot.sim_iid}
\title{Plot simulated random intercepts}
\usage{
\method{plot}{sim_iid}(x, y, ..., link = c("identity", "log", "logit"),
  baseline, center = c("mean", "bottom", "top"), quantiles = c(0.025,
  0.1, 0.5, 0.9, 0.975))
}
\arguments{
\item{x}{a \code{sim_iid} object. Which is the output of \code{\link{simulate_iid}}}

\item{y}{currently ignored}

\item{...}{currently ignored}

\item{link}{which link to use for back transformation}

\item{baseline}{optional baseline for the time series}

\item{center}{defines how to center the time series to the baseline. Options are: \code{start} all time series start at the baseline; \code{mean} the average of the time series is the baseline; \code{bottom} the lowest value of the time series equals the baseline; \code{top} the highest value of the time series equals the baseline}

\item{quantiles}{which quantiles are shown on the plot}
}
\value{
a ggplot2 object
}
\description{
Plot simulated random intercepts
}
\examples{
set.seed(20181202)
x <- simulate_iid(sigma = 0.25)
plot(x)
plot(x, link = "log")
plot(x, link = "logit")
}
\seealso{
Other priors: \code{\link{plot.sim_rw}},
  \code{\link{select_change}},
  \code{\link{select_divergence}},
  \code{\link{select_poly}}, \code{\link{select_quantile}},
  \code{\link{simulate_iid}}, \code{\link{simulate_rw}}
}
\concept{priors}
