% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_rw.R
\name{plot.sim_rw}
\alias{plot.sim_rw}
\title{Plot simulated random walks}
\usage{
\method{plot}{sim_rw}(x, y, ..., type = c("all", "divergence",
  "stationary", "quantile", "change"), link = c("identity", "log",
  "logit"))
}
\arguments{
\item{x}{a \code{sim_rw} object. Which is the output of  \code{\link{simulate_rw}}}

\item{y}{currently ignored}

\item{...}{currently ignored}

\item{type}{which plot to create. \code{"all"} displays all simulations. \code{"divergence"} displays the most divergent simulations. \code{"stationary"} displays the smimulations with the smallest differences for the reference. \code{"quantile"} displays the enveloppes around the simulations. \code{"changes"} displays the simulations with the highest number of changes in directions in the random walk.}

\item{link}{which link to use for back transformation}
}
\value{
a ggplot2 object
}
\description{
Plot simulated random walks
}
\seealso{
Other priors: \code{\link{simulate_rw}}
}
\concept{priors}
