% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inla_methods.R
\name{prec.inla}
\alias{prec.inla}
\title{Define the prec method for an inla output object}
\usage{
\method{prec}{inla}(model, ...)
}
\arguments{
\item{model}{an inla output}

\item{...}{used to pass the 'prior' argument,
as logical (default is TRUE) to indicate if it is
to retrieve the prior or the posterior precision.}
}
\description{
Define the prec method for an inla output object
}
\details{
extract the joint prior precision for the
latent field at the mode of the hyperparameters
}
