% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rgeneric.R
\name{rgeneric}
\alias{rgeneric}
\alias{rgeneric.default}
\title{Organize data for the latent GMRF R interface for \code{INLA}.}
\usage{
rgeneric(model, n, debug = FALSE, ...)

\method{rgeneric}{default}(model, n, debug = FALSE, ...)
}
\arguments{
\item{model}{an object used to define the model.
See the 'rgeneric' vignette from the INLA package.}

\item{n}{integer with the dimension of the model}

\item{debug}{logical indicating debug state.}

\item{...}{additional arguments to be used internally
for the model, for example, additional data.}
}
\value{
\code{rgeneric}/ \code{inla.rgeneric} object.
}
\description{
Organize data for the latent GMRF R interface for \code{INLA}.

The rgeneric default method.
}
