% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zeroinflated_nbinomial.R
\name{rzanbinom}
\alias{rzanbinom}
\title{Zero altered negative binomial}
\usage{
rzanbinom(n, mu, size, prob, tol = 2e-10)
}
\arguments{
\item{n}{number of random values to return.}

\item{mu}{alternative parametrization via mean: see \sQuote{Details}.}

\item{size}{target for number of successful trials, or dispersion
    parameter (the shape parameter of the gamma mixing distribution).
    Must be strictly positive, need not be integer.}

\item{prob}{the point mass of zero}

\item{tol}{the tolerance for low numbers}
}
\description{
Zero altered negative binomial
}
