% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zeroinflated_poisson.R
\name{rzapois}
\alias{rzapois}
\title{Zero-altered Poisson}
\usage{
rzapois(n, lambda, prob, tol = 2e-10)
}
\arguments{
\item{prob}{the point mass of zero}

\item{tol}{the tolerance for low numbers}
}
\description{
Generate random numbers from a zero-altered Poisson distribution
}
