% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zeroinflated_nbinomial.R
\name{rzinbinom}
\alias{rzinbinom}
\title{Zero inflated negative binomial}
\usage{
rzinbinom(n, mu, size, prob)
}
\arguments{
\item{n}{number of random values to return.}

\item{mu}{alternative parametrization via mean: see \sQuote{Details}.}

\item{size}{target for number of successful trials, or dispersion
    parameter (the shape parameter of the gamma mixing distribution).
    Must be strictly positive, need not be integer.}

\item{prob}{the mass of extra zero's}
}
\description{
Zero inflated negative binomial
}
