% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_rw.R
\name{select_divergence}
\alias{select_divergence}
\title{Select Diverging Simulations from an \code{sim_rw} Object}
\usage{
select_divergence(x, n = 10)
}
\arguments{
\item{x}{An \code{sim_rw} object.
Which is the output of \verb{\link{simulate_rw}}}

\item{n}{The number of simulations to plot when only a subset is shown.}
}
\description{
The selection will contain the most extreme simulations base on either the
minimum effect or the maximum effect within the simulation.
}
\seealso{
Other priors: 
\code{\link{plot.sim_iid}()},
\code{\link{plot.sim_rw}()},
\code{\link{select_change}()},
\code{\link{select_poly}()},
\code{\link{select_quantile}()},
\code{\link{simulate_iid}()},
\code{\link{simulate_rw}()}
}
\concept{priors}
