% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_rw.R
\name{select_poly}
\alias{select_poly}
\title{Select random walks best matching some polygon coefficients}
\usage{
select_poly(x, coefs = c(0, -1), n = 10)
}
\arguments{
\item{x}{a \code{sim_rw} object. Which is the output of  \code{\link{simulate_rw}}}

\item{coefs}{the polynomial coefficients}

\item{n}{the number of simulations to plot when only a subset is shown.}
}
\description{
The target coefficients will be rescaled to have norm 1. The coefficients of the simulations will be rescaled by the largest norm over all simulations.
}
\seealso{
Other priors: \code{\link{plot.sim_iid}},
  \code{\link{plot.sim_rw}}, \code{\link{select_change}},
  \code{\link{select_divergence}},
  \code{\link{select_quantile}},
  \code{\link{simulate_iid}}, \code{\link{simulate_rw}}
}
\concept{priors}
