% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_rw.R
\name{simulate_rw}
\alias{simulate_rw}
\title{simulate data from a second order random walk}
\usage{
simulate_rw(sigma = NULL, tau = NULL, length = 10, start = 1,
  order = 1, n_sim = 1000)
}
\arguments{
\item{sigma}{the standard deviation of the random walk process}

\item{tau}{the precision of the random walk process}

\item{length}{the length of the time series}

\item{start}{the starting values of the time series}

\item{order}{1 for first order random walk or 2 for second order random walk}

\item{n_sim}{the number of simulations}
}
\value{
a data.frame with simulated time series from the random walk
}
\description{
simulate data from a second order random walk
}
\examples{
set.seed(20181202)
x <- simulate_rw(sigma = 0.1, start = -10, length = 40)
head(x)
y <- simulate_rw(sigma = 0.001, start = -10, length = 40, order = 2)
head(y)
}
\seealso{
Other priors: \code{\link{plot.sim_iid}},
  \code{\link{plot.sim_rw}}, \code{\link{simulate_iid}}
}
\concept{priors}
