% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_rw2.R
\name{simulate_rw2}
\alias{simulate_rw2}
\title{simulate data from a second order random walk}
\usage{
simulate_rw2(sigma = 0.01, tau = NULL, length = 10, start = 1,
  n_sim = 1000)
}
\arguments{
\item{sigma}{the standard deviation of the random walk process}

\item{tau}{the precision of the random walk process}

\item{length}{the length of the time series}

\item{start}{the starting values of the time series}

\item{n_sim}{the number of simulations}
}
\value{
a data.frame with simulated time series from the second order random walk
}
\description{
simulate data from a second order random walk
}
