% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_dispersion.R
\docType{methods}
\name{test_dispersion}
\alias{test_dispersion}
\alias{test_dispersion,inla-method}
\title{Use simulations to test for overdispersion or underdispersion}
\usage{
test_dispersion(object, nsim = 1000, plot = TRUE)

\S4method{test_dispersion}{inla}(object, nsim = 1000, plot = TRUE)
}
\arguments{
\item{object}{the INLA model}

\item{nsim}{the number of simulation}

\item{plot}{display a plot of of the simulated values}
}
\description{
Use simulations to test for overdispersion or underdispersion
}
