% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_distribution.R
\docType{methods}
\name{test_distribution}
\alias{test_distribution}
\alias{test_distribution,inla-method}
\title{Use simulations to compare the observed distribution with the modelled distribution}
\usage{
test_distribution(object, nsim = 1000, plot = TRUE)

\S4method{test_distribution}{inla}(object, nsim = 1000, plot = TRUE)
}
\arguments{
\item{object}{the INLA model}

\item{nsim}{the number of simulation}

\item{plot}{display a plot of of the simulated values}
}
\description{
Use simulations to compare the observed distribution with the modelled distribution
}
