% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residuals.R
\name{var_nbinom}
\alias{var_nbinom}
\alias{var_zinbinom1}
\alias{var_zipois1}
\title{Variance of the negative binomial distribution}
\usage{
var_nbinom(mu, size)

var_zinbinom1(mu, size, zero)

var_zipois1(mu, zero)
}
\arguments{
\item{mu}{mean of the distribution.
Must be on the original scale.}

\item{size}{Size of the negative binomial distribution.
Must be strict positive.}

\item{zero}{Probability of the point mass at zero.}
}
\description{
The type 1 zero-inflated negative binomial distribution is a standard
negative binomial distribution with an additional point mass at zero.

The type 1 zero-inflated poisson is a standard Poisson distribution with an
additional point mass at zero.
}
\concept{variance}
