% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa_classes.R, R/cgeneric.R
\docType{class}
\name{cgeneric-class}
\alias{cgeneric-class}
\alias{cgeneric.cgeneric}
\alias{cgeneric.inla.cgeneric}
\alias{mapper1}
\alias{print.cgeneric}
\alias{summary.cgeneric}
\alias{plot.cgeneric}
\title{Organize data for the latent GMRF C interface for \code{INLA}.}
\usage{
\method{cgeneric}{cgeneric}(model, ...)

\method{cgeneric}{inla.cgeneric}(model, ...)

mapper1(model)

\method{print}{cgeneric}(x, ...)

\method{summary}{cgeneric}(object, ...)

\method{plot}{cgeneric}(x, y, ...)
}
\arguments{
\item{...}{not used}

\item{x}{a cgeneric object}

\item{object}{a cgeneric object}

\item{y}{not used}
}
\description{
Organize data for the latent GMRF C interface for \code{INLA}.
}
\section{Functions}{
\itemize{
\item \code{cgeneric(cgeneric)}: Returns the model object unchanged.

\item \code{cgeneric(inla.cgeneric)}: Converts a regular \code{inla.cgeneric} object to \code{cgeneric}.

\item \code{mapper1()}: A default mapper for a cgeneric/rgeneric model

\item \code{print(cgeneric)}: Print the cgeneric object

\item \code{summary(cgeneric)}: A summary for a cgeneric object

\item \code{plot(cgeneric)}: A plot for a cgeneric object

}}
