% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa_classes.R, R/cgeneric.R
\docType{class}
\name{cgeneric-class}
\alias{cgeneric-class}
\alias{cgeneric}
\alias{cgeneric.character}
\alias{cgeneric.function}
\alias{cgenericBuilder}
\alias{cgeneric_libpath}
\title{The \code{cgeneric} class organize data needed
to work with Gaussian Markov Random Fields - GMRF,
defined as a C interface for \code{INLA}.}
\usage{
cgeneric(model, ...)

\method{cgeneric}{character}(model, ...)

\method{cgeneric}{`function`}(model, ...)

cgenericBuilder(...)

cgeneric_libpath(debug, package, useINLAprecomp)
}
\arguments{
\item{model}{object class for what a \code{cgeneric} method exists.
E.g., if it is a character, a specific function will be called.
E.g. cgeneric("iid", ...") calls cgeneric_iid(...).}

\item{...}{arguments passed from the
\code{\link[=cgeneric]{cgeneric()}} methods.}

\item{debug}{integer, used as verbose in debug.}

\item{package}{character giving the name of the package
that contains the \code{cgeneric} model.}

\item{useINLAprecomp}{logical, indicating if it is to use
the shared object previously copied and compiled by INLA.}
}
\value{
named list of \code{cgeneric} class containing
the named list \code{f} that contain \code{model} (a character
always equal to \code{cgeneric}), \code{n} (integer)
and \code{cgeneric} as a named list that contains the
data needed to define the model. Each element on
...$f$cgeneric is also a named list containing
\code{ints}, \code{doubles}, \code{characters}, \code{matrices}
and \code{smatrices}.
cgeneric_libpath returns character with the path to the shared lib.
}
\description{
A GMRF is defined from model parameters \eqn{\theta} that
would parametrize a (sparse) precision matrix.

The elements of a GMR are:
\itemize{
\item \code{graph} to define the non-zero precision matrix pattern.
only the upper triangle including the diagonal is needed.
\item \code{Q} vector where the
\itemize{
\item first element (N) is the size of the matrix,
\item second element (M) is the number of non-zero
elements in the upper part (including) diagonal
\item the remaining (M) elements are the actual
precision (upper triangle plus diagonal) elements
whose order shall follow the graph definition.
}
\item \code{mu} the mean vector,
\item \code{initial} vector with
\itemize{
\item first element as the number of the parameters in the model
\item remaining elements should be the initials for the model parameters.
}
\item \code{log.norm.const} log of the normalizing constant.
\item \code{log.prior} log of the prior for the model parameters.
}
}
\note{
The \code{graph} and \code{Q} non-zero pattern should match,
its elements should be ordered by row,
and only its upper part stored.
}
\seealso{
\code{\link[INLA:cgeneric]{INLA::cgeneric()}} and \code{\link[=methods]{methods()}}
}
