% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa_classes.R, R/cgeneric.R
\docType{class}
\name{cgeneric-class}
\alias{cgeneric-class}
\alias{cgeneric.character}
\title{A \code{cgeneric} model described in \code{\link[=cgeneric]{cgeneric()}}.}
\usage{
\method{cgeneric}{character}(model, ...)
}
\arguments{
\item{model}{object class for what a \code{cgeneric} method exists.
E.g., if it is a character, a specific function will be called:
cgeneric("iid", ...") calls cgeneric_iid(...)}

\item{...}{additional arguments passed to to methods.
Some arguments can be used to define specific behavior,
such as \code{debug} (integer, used as verbose in debug),
\code{useINLAprecomp} (logical, indicating if it is to use
the shared object previously copied and compiled by INLA),
\code{package} (character used if \code{useINLAprecomp} is TRUE, with
the package name to build the path) and \code{libpath} (character,
with the path to the shared dynamic library object: this
override \code{useINLAprecomp} and \code{package}).}
}
\description{
A \code{cgeneric} model described in \code{\link[=cgeneric]{cgeneric()}}.
}
\details{
Method for when \code{model} is a character.
E.g. cgeneric(model = "generic0")
calls \link{cgeneric_generic0}
}
\section{Functions}{
\itemize{
\item \code{cgeneric(character)}: This calls \code{\link[INLA:cgeneric]{INLA::inla.cgeneric.define()}}

}}
\examples{
cg <- cgeneric(model = "fake", n = 5, package = "INLAtools")
str(cg)
}
