% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dispersion_check.R
\name{dispersion}
\alias{dispersion}
\title{Calculate a measure for dispersion
The measure is calculated as the average of the squared Pearson residuals}
\usage{
dispersion(observed, fitted, variance)
}
\arguments{
\item{observed}{the observed values}

\item{fitted}{the fitted values}

\item{variance}{the variance of the fitted values}
}
\description{
Calculate a measure for dispersion
The measure is calculated as the average of the squared Pearson residuals
}
\examples{
library(INLA)
set.seed(20181202)
model <- inla(
  poisson ~ 1,
  family = "poisson",
  data = data.frame(
    poisson = rpois(20, lambda = 10),
    base = 1
  ),
  control.predictor = list(compute = TRUE)
)
dispersion(
  observed = get_observed(model),
  fitted = fitted(model),
  variance = fitted(model)
)
}
\seealso{
Other statistics: \code{\link{fitted,inla-method}},
  \code{\link{get_observed}},
  \code{\link{residuals,inla-method}}
}
\concept{statistics}
