% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_dispersion.R
\name{dispersion}
\alias{dispersion}
\title{Calculate a measure for dispersion
The measure is calculated as the average of the squared Pearson residuals}
\usage{
dispersion(observed, fitted, variance)
}
\arguments{
\item{observed}{the observed values}

\item{fitted}{the fitted values}

\item{variance}{the variance of the fitted values}
}
\description{
Calculate a measure for dispersion
The measure is calculated as the average of the squared Pearson residuals
}
