% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dispersion_check.R
\docType{methods}
\name{dispersion_check}
\alias{dispersion_check}
\alias{dispersion_check,inla-method}
\title{Use simulations to check for overdispersion or underdispersion}
\usage{
dispersion_check(object, nsim = 1000)

\S4method{dispersion_check}{inla}(object, nsim = 1000)
}
\arguments{
\item{object}{the INLA model}

\item{nsim}{the number of simulation}
}
\description{
Use simulations to check for overdispersion or underdispersion
}
\examples{
library(INLA)
set.seed(20181202)
model <- inla(
  poisson ~ 1,
  family = "poisson",
  data = data.frame(
    poisson = rpois(20, lambda = 10),
    base = 1
  ),
  control.predictor = list(compute = TRUE)
)
dc <- dispersion_check(model)
str(dc)
}
\seealso{
Other checks: \code{\link{distribution_check}},
  \code{\link{fast_distribution_check}}
}
\concept{checks}
