% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution_check.R
\docType{methods}
\name{distribution_check}
\alias{distribution_check}
\alias{distribution_check,inla-method}
\title{Use simulations to compare the observed distribution with the modelled
distribution}
\usage{
distribution_check(object, nsim = 1000, seed = 0L)

\S4method{distribution_check}{inla}(object, nsim = 1000, seed = 0L)
}
\arguments{
\item{object}{the INLA model}

\item{nsim}{the number of simulation}

\item{seed}{See the same argument in \code{?inla.qsample} for further
information. In order to produce reproducible results, you ALSO need to make
sure the RNG in R is in the same state, see example below.  When \code{seed}
is non-zero, \code{num.threads} is forced to "1:1" and parallel.configs is
set to \code{FALSE}, since parallel sampling would not produce a
reproducible sequence of pseudo-random numbers.}
}
\description{
Use simulations to compare the observed distribution with the modelled
distribution
}
\examples{
\donttest{
library(INLA)
model <- inla(
  poisson ~ 1,
  family = "poisson",
  data = data.frame(
    poisson = rpois(20, lambda = 10),
    base = 1
  ),
  control.predictor = list(compute = TRUE),
  control.compute = list(config = TRUE)
)
distribution_check(model, seed = 20181202)
}
}
\seealso{
Other checks: 
\code{\link{dispersion_check}()},
\code{\link{fast_aggregation_check}()},
\code{\link{fast_aggregation_check,inla-method}},
\code{\link{get_anomaly}()}
}
\concept{checks}
