% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution_check.R
\docType{methods}
\name{distribution_check}
\alias{distribution_check}
\alias{distribution_check,inla-method}
\title{Use simulations to compare the observed distribution with the modelled distribution}
\usage{
distribution_check(object, nsim = 1000, plot = TRUE)

\S4method{distribution_check}{inla}(object, nsim = 1000, plot = TRUE)
}
\arguments{
\item{object}{the INLA model}

\item{nsim}{the number of simulation}

\item{plot}{display a plot of of the simulated values}
}
\description{
Use simulations to compare the observed distribution with the modelled distribution
}
