% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution_check.R
\docType{methods}
\name{distribution_check}
\alias{distribution_check}
\alias{distribution_check,inla-method}
\title{Use simulations to compare the observed distribution with the modelled
distribution}
\usage{
distribution_check(object, nsim = 1000)

\S4method{distribution_check}{inla}(object, nsim = 1000)
}
\arguments{
\item{object}{the INLA model}

\item{nsim}{the number of simulation}
}
\description{
Use simulations to compare the observed distribution with the modelled
distribution
}
\examples{
\donttest{
library(INLA)
set.seed(20181202)
model <- inla(
  poisson ~ 1,
  family = "poisson",
  data = data.frame(
    poisson = rpois(20, lambda = 10),
    base = 1
  ),
  control.predictor = list(compute = TRUE),
  control.compute = list(config = TRUE)
)
distribution_check(model)
}
}
\seealso{
Other checks: 
\code{\link{dispersion_check}()},
\code{\link{fast_distribution_check}()}
}
\concept{checks}
