% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fast_distribution_check.R
\docType{methods}
\name{fast_distribution_check}
\alias{fast_distribution_check}
\alias{fast_distribution_check,inla-method}
\title{Use simulations to compare the observed distribution with the modelled distribution}
\usage{
fast_distribution_check(object, nsim = 1000, plot = TRUE)

\S4method{fast_distribution_check}{inla}(object, nsim = 1000,
  plot = TRUE)
}
\arguments{
\item{object}{the INLA model}

\item{nsim}{the number of simulation}

\item{plot}{display a plot of of the simulated values}
}
\description{
This check uses the fitted values and thus ignores the uncertainty on the predictions
}
