% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_observed.R
\docType{methods}
\name{fitted,inla-method}
\alias{fitted,inla-method}
\title{Extract the fitted values from an INLA model}
\usage{
\S4method{fitted}{inla}(object, ...)
}
\arguments{
\item{object}{an object for which the extraction of model fitted values is
    meaningful.}

\item{...}{other arguments.}
}
\description{
Extract the fitted values from an INLA model
}
\examples{
library(INLA)
set.seed(20181202)
model <- inla(#'   poisson ~ 1,
  poisson ~ 1,
  family = "poisson",
  data = data.frame(
    poisson = rpois(20, lambda = 10),
    base = 1
  ),
  control.predictor = list(compute = TRUE)
)
fitted(model)
}
\seealso{
Other statistics: \code{\link{dispersion}},
  \code{\link{get_observed}},
  \code{\link{residuals,inla-method}}
}
\concept{statistics}
