% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_observed.R
\name{fitted}
\alias{fitted}
\title{get the mean of the fitted values from the model object}
\usage{
fitted(object)
}
\arguments{
\item{object}{the INLA model}
}
\description{
get the mean of the fitted values from the model object
}
