% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_data.R
\name{generate_data}
\alias{generate_data}
\title{Generate dummy data with several distributions}
\usage{
generate_data(
  a = 0,
  sigma_random = 0.5,
  n_random = 20,
  n_replicate = 10,
  nb_size = 1,
  b_size = 5,
  zero_inflation = 0.5
)
}
\arguments{
\item{a}{the intercept of the latent variable}

\item{sigma_random}{The standard error for the random effect \eqn{\sigma_r}.}

\item{n_random}{the number of random effect levels (groups).}

\item{n_replicate}{the number of observation per random effect level.}

\item{nb_size}{the size parameter of the negative binomial distribution.
Passed to the \code{size} parameter of \code{stats::rnbinom()}.}

\item{b_size}{the size parameter of the binomial distribution. Passed to the
\code{size} parameter of \code{stats::rbinom()}.}

\item{zero_inflation}{the probability the the observed value stems for the a
point mass in zero.}
}
\value{
A \code{data.frame}
\itemize{
\item \code{ìd} the id of the random effect.
\item \code{eta} the latent variable.
\item \code{zero_inflation} use the point mass in zero.
\item \code{poisson} the Poisson distributed variable.
\item \code{zipoisson} the zero-inflated Poisson distributed variable.
\item \code{negbin} the negative binomial distributed variable.
\item \code{zinegbin} the zero-inflated negative binomial distributed variable.
\item \code{binom} the binomial distributed variable.
}
}
\description{
All distributions share the same latent variable \eqn{\eta_{ij} = a + b_i}
with \eqn{b_i = N(0, \sigma_r)}.
}
\details{
\itemize{
\item The Poisson distribution uses \eqn{\lambda = e^{\eta_{ij}}}.
\item The negation binomial distribution uses \eqn{\mu = e^{\eta_{ij}}}.
\item The binomial distribution uses
\eqn{\pi_{ij} = e^{\eta_{ij}}/(e^{\eta_{ij}}+ 1)}.
}
}
\examples{
set.seed(20181202)
head(generate_data())
}
\seealso{
Other utils: 
\code{\link{plot.dispersion_check}()},
\code{\link{plot.distribution_check}()}
}
\concept{utils}
