% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_observed.R
\name{get_observed}
\alias{get_observed}
\title{get the observed values from the model object}
\usage{
get_observed(model)
}
\arguments{
\item{model}{the INLA model}
}
\description{
get the observed values from the model object
}
