% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kronecker.R
\name{kronecker}
\alias{kronecker}
\alias{kronecker,cgeneric,cgeneric-method}
\alias{kronecker,cgeneric,rgeneric-method}
\alias{kronecker,rgeneric,cgeneric-method}
\alias{kronecker,rgeneric,rgeneric-method}
\title{Kronecker between \code{cgeneric}|\code{rgeneric} to implement interaction
between GMRF models.}
\usage{
\S4method{kronecker}{cgeneric,cgeneric}(X, Y, FUN = "*", make.dimnames = FALSE, ...)

\S4method{kronecker}{cgeneric,rgeneric}(X, Y, FUN = "*", make.dimnames = FALSE, ...)

\S4method{kronecker}{rgeneric,cgeneric}(X, Y, FUN = "*", make.dimnames = FALSE, ...)

\S4method{kronecker}{rgeneric,rgeneric}(X, Y, FUN = "*", make.dimnames = FALSE, ...)
}
\arguments{
\item{X}{\code{cgeneric} or \code{rgeneric}}

\item{Y}{\code{cgeneric} or \code{rgeneric}}

\item{FUN}{see \code{\link[=kronecker]{kronecker()}}}

\item{make.dimnames}{see \code{\link[=kronecker]{kronecker()}}}

\item{...}{see \code{\link[=kronecker]{kronecker()}}}
}
\value{
A \code{cgeneric} object if kronecker between two \code{cgeneric},
otherwise a \code{rgeneric} object.
}
\description{
Kronecker between \code{cgeneric}|\code{rgeneric} to implement interaction
between GMRF models.
}
\section{Functions}{
\itemize{
\item \code{kronecker(X = cgeneric, Y = cgeneric)}: Build a \code{cgeneric} to implement the interaction between
two GMRF models each one implemented as a \code{cgeneric}.

\item \code{kronecker(X = cgeneric, Y = rgeneric)}: Build a \code{rgeneric} to implement the interaction between two GMRF
models, the first as a \code{cgeneric} and the second as a \code{rgeneric}.

\item \code{kronecker(X = rgeneric, Y = cgeneric)}: Build a \code{rgeneric} to implement the interaction between two GMRF
models, the first as a \code{rgeneric} and the second as a \code{cgeneric}.

\item \code{kronecker(X = rgeneric, Y = rgeneric)}: Build a \code{rgeneric} to implement the interaction between
two GMRF models each one implemented as a \code{rgeneric}.

}}
