% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_distribution_check.R
\name{plot.dispersion_check}
\alias{plot.dispersion_check}
\title{Plot the results from a dispersion check}
\usage{
\method{plot}{dispersion_check}(x, y, ...)
}
\arguments{
\item{x}{a dispersion_check object. Which is the output of
\verb{\link{dispersion_check}}}

\item{y}{currently ignored}

\item{...}{currently ignored}
}
\value{
a ggplot2 object
}
\description{
Plot the results from a dispersion check
}
\examples{
library(INLA)
set.seed(20181202)
model <- inla(
  poisson ~ 1,
  family = "poisson",
  data = data.frame(
    poisson = rpois(20, lambda = 10),
    base = 1
  ),
  control.predictor = list(compute = TRUE)
)
dc <- dispersion_check(model)
plot(dc)
}
\seealso{
Other utils: 
\code{\link{generate_data}()},
\code{\link{plot.distribution_check}()}
}
\concept{utils}
