% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_distribution_check.R
\name{plot.distribution_check}
\alias{plot.distribution_check}
\title{Plot the results from a distribution check}
\usage{
\method{plot}{distribution_check}(x, y, ..., n = FALSE, scales = "fixed")
}
\arguments{
\item{x}{a distribution_check object. Which is the output of
\verb{\link{fast_distribution_check}}}

\item{y}{currently ignored}

\item{...}{currently ignored}

\item{n}{display the number of observations}

\item{scales}{Should scales be fixed (\code{"fixed"}, the default),
free (\code{"free"}), or free in one dimension (\code{"free_x"},
\code{"free_y"})?}
}
\value{
A \verb{\link[ggplot2]\{ggplot\}} object.
}
\description{
Plot the results from a distribution check
}
\examples{
library(INLA)
set.seed(20181202)
model <- inla(
  poisson ~ 1,
  family = "poisson",
  data = data.frame(
    poisson = rpois(20, lambda = 10),
    base = 1
  ),
  control.predictor = list(compute = TRUE)
)
fdc <- fast_distribution_check(model)
plot(fdc)
}
\seealso{
Other utils: 
\code{\link{generate_data}()},
\code{\link{plot.dispersion_check}()}
}
\concept{utils}
