% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_idd.R
\name{plot.sim_iid}
\alias{plot.sim_iid}
\title{Plot Simulated Random Intercepts}
\usage{
\method{plot}{sim_iid}(
  x,
  y,
  ...,
  link = c("identity", "log", "logit"),
  baseline,
  center = c("mean", "bottom", "top"),
  quantiles = c(0.025, 0.1, 0.5, 0.9, 0.975)
)
}
\arguments{
\item{x}{A \code{sim_iid} object. Which is the output of \verb{\link{simulate_iid}}.}

\item{y}{Currently ignored.}

\item{...}{Currently ignored.}

\item{link}{Which link to use for back transformation.}

\item{baseline}{Optional baseline for the time series.}

\item{center}{Defines how to center the time series to the baseline.
Options are: \code{start} all time series start at the baseline;
\code{mean} the average of the time series is the baseline;
\code{bottom} the lowest value of the time series equals the baseline;
\code{top} the highest value of the time series equals the baseline.}

\item{quantiles}{Which quantiles are shown on the plot.}
}
\value{
A \verb{\link[ggplot2]\{ggplot\}} object.
}
\description{
Plot Simulated Random Intercepts
}
\examples{
set.seed(20181202)
x <- simulate_iid(sigma = 0.25)
plot(x)
plot(x, link = "log")
plot(x, link = "logit")
}
\seealso{
Other priors: 
\code{\link{plot.sim_rw}()},
\code{\link{select_change}()},
\code{\link{select_divergence}()},
\code{\link{select_poly}()},
\code{\link{select_quantile}()},
\code{\link{simulate_iid}()},
\code{\link{simulate_rw}()}
}
\concept{priors}
