% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_idd.R
\name{plot.sim_iid}
\alias{plot.sim_iid}
\title{Plot simulated random intercepts}
\usage{
\method{plot}{sim_iid}(x, y, ..., link = c("identity", "log", "logit"))
}
\arguments{
\item{x}{a \code{sim_iid} object. Which is the output of \code{\link{simulate_iid}}}

\item{y}{currently ignored}

\item{...}{currently ignored}

\item{link}{which link to use for back transformation}
}
\value{
a ggplot2 object
}
\description{
Plot simulated random intercepts
}
\examples{
set.seed(20181202)
x <- simulate_iid(sigma = 0.25)
plot(x)
plot(x, link = "log")
plot(x, link = "logit")
}
\seealso{
Other priors: \code{\link{plot.sim_rw}},
  \code{\link{simulate_iid}}, \code{\link{simulate_rw}}
}
\concept{priors}
