% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_rw.R
\name{plot.sim_rw}
\alias{plot.sim_rw}
\title{Plot simulated random walks}
\usage{
\method{plot}{sim_rw}(x, y, ..., link = c("identity", "log", "logit"),
  baseline)
}
\arguments{
\item{x}{a \code{sim_rw} object. Which is the output of  \code{\link{simulate_rw}}}

\item{y}{currently ignored}

\item{...}{currently ignored}

\item{link}{which link to use for back transformation}

\item{baseline}{optional baseline for the time series}
}
\value{
a \code{\link[ggplot2]{ggplot}} object
}
\description{
Plot simulated random walks
}
\examples{
\donttest{
set.seed(20181202)
x <- simulate_rw(sigma = 0.05, start = -10, length = 40)
plot(x)
plot(select_quantile(x))
plot(select_quantile(x), link = "log")
plot(select_quantile(x), link = "logit")
x <- simulate_rw(sigma = 0.001, start = -10, length = 40, order = 2)
plot(x)
plot(select_quantile(x))
plot(select_quantile(x), link = "log")
plot(select_quantile(x), link = "logit")
}
}
\seealso{
Other priors: \code{\link{plot.sim_iid}},
  \code{\link{select_change}},
  \code{\link{select_divergence}},
  \code{\link{select_poly}}, \code{\link{select_quantile}},
  \code{\link{simulate_iid}}, \code{\link{simulate_rw}}
}
\concept{priors}
