% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_rw.R
\name{plot.sim_rw}
\alias{plot.sim_rw}
\title{Plot simulated random walks}
\usage{
\method{plot}{sim_rw}(x, y, ..., type = c("all", "divergence",
  "stationary", "quantile", "change", "poly"), link = c("identity",
  "log", "logit"), n = 10, coefs = c(0, -1))
}
\arguments{
\item{x}{a \code{sim_rw} object. Which is the output of  \code{\link{simulate_rw}}}

\item{y}{currently ignored}

\item{...}{currently ignored}

\item{type}{which plot to create. \code{"all"} displays all simulations. \code{"divergence"} displays the most divergent simulations. \code{"stationary"} displays the smimulations with the smallest differences for the reference. \code{"quantile"} displays the enveloppes around the simulations. \code{"changes"} displays the simulations with the highest number of changes in directions in the random walk.}

\item{link}{which link to use for back transformation}

\item{n}{the number of simulations to plot when only a subset is shown.}

\item{coefs}{the polynomial coefficients}
}
\value{
a ggplot2 object
}
\description{
Plot simulated random walks
}
\examples{
\donttest{
set.seed(20181202)
x <- simulate_rw(sigma = 0.05, start = -10, length = 40)
plot(x)
plot(x, type = "quantile")
plot(x, type = "divergence")
plot(x, type = "stationary")
plot(x, type = "change")
plot(x, type = "quantile", link = "log")
plot(x, type = "quantile", link = "logit")
y <- simulate_rw(sigma = 0.001, start = -10, length = 40, order = 2)
plot(y)
plot(y, type = "quantile")
plot(y, type = "divergence")
plot(y, type = "stationary")
plot(y, type = "change")
plot(y, type = "quantile", link = "log")
plot(y, type = "quantile", link = "logit")
}
}
\seealso{
Other priors: \code{\link{plot.sim_iid}},
  \code{\link{simulate_iid}}, \code{\link{simulate_rw}}
}
\concept{priors}
