% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_rw.R
\name{plot.sim_rw}
\alias{plot.sim_rw}
\title{Plot Simulated Random Walks}
\usage{
\method{plot}{sim_rw}(x, y, ..., link = c("identity", "log", "logit"),
  baseline, center = c("start", "mean", "bottom", "top"))
}
\arguments{
\item{x}{An \code{sim_rw} object.
Which is the output of  \code{\link{simulate_rw}}}

\item{y}{Currently ignored.}

\item{...}{Currently ignored.}

\item{link}{Which link to use for back transformation.}

\item{baseline}{Optional baseline for the time series.}

\item{center}{Defines how to center the time series to the baseline.
Options are: \code{start} all time series start at the baseline;
\code{mean} the average of the time series is the baseline;
\code{bottom} the lowest value of the time series equals the baseline;
\code{top} the highest value of the time series equals the baseline.}
}
\value{
A \code{\link[ggplot2]{ggplot}} object.
}
\description{
Plot Simulated Random Walks
}
\examples{
\donttest{
set.seed(20181202)
x <- simulate_rw(sigma = 0.05, start = -10, length = 40)
plot(x)
plot(select_quantile(x))
plot(select_quantile(x), link = "log")
plot(select_quantile(x), link = "logit")
x <- simulate_rw(sigma = 0.001, start = -10, length = 40, order = 2)
plot(x)
plot(select_quantile(x))
plot(select_quantile(x), link = "log")
plot(select_quantile(x), link = "logit")
}
}
\seealso{
Other priors: \code{\link{plot.sim_iid}},
  \code{\link{select_change}},
  \code{\link{select_divergence}},
  \code{\link{select_poly}}, \code{\link{select_quantile}},
  \code{\link{simulate_iid}}, \code{\link{simulate_rw}}
}
\concept{priors}
