% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prec2sd.R
\name{prec2sd}
\alias{prec2sd}
\title{Convert the posterior marginal of a precision to a standard deviation}
\usage{
prec2sd(marg)
}
\arguments{
\item{marg}{A matrix with columns "y" and "x" where "y" is the marginal of
the precision.}
}
\value{
A \code{data.frame} with the mean, standard deviation and 2.5\%, 25\%, 50\%,
75\% and 97.5\% quantiles of the posterior of the standard deviation.
}
\description{
Convert the posterior marginal of a precision to a standard deviation
}
\examples{
stopifnot(require(INLA))
model <- inla(Sepal.Length ~ Species, data = iris, family = "gaussian")
marg <- model$marginals.hyperpar[["Precision for the Gaussian observations"]]
prec2sd(marg)
}
