% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residuals.R
\docType{methods}
\name{residuals,inla-method}
\alias{residuals,inla-method}
\title{Calculate the residuals from an INLA model}
\usage{
\S4method{residuals}{inla}(object, type = c("pearson", "deviance",
  "working", "response", "partial"), ...)
}
\arguments{
\item{object}{the INLA object}

\item{type}{the type of residuals which should be returned.
    The alternatives are: \code{"deviance"} (default), \code{"pearson"},
    \code{"working"}, \code{"response"}, and \code{"partial"}.
    Can be abbreviated.}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Calculate the residuals from an INLA model
}
