% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residuals.R
\name{residuals,inla-method}
\alias{residuals,inla-method}
\title{Calculate the Residuals From an INLA Model}
\usage{
\S4method{residuals}{inla}(
  object,
  type = c("pearson", "deviance", "working", "response", "partial"),
  ...
)
}
\arguments{
\item{object}{An \code{inla} object.}

\item{type}{Currently only Pearson residuals are available.
Other types are only listed for compatibility with the default \code{residuals}
function.}

\item{...}{Currently ignored.}
}
\description{
Calculate the Residuals From an INLA Model
}
\examples{
library(INLA)
set.seed(20181202)
model <- inla(
  poisson ~ 1,
  family = "poisson",
  data = data.frame(
    poisson = rpois(20, lambda = 10),
    base = 1
  ),
  control.predictor = list(compute = TRUE)
)
residuals(model)
}
\seealso{
Other statistics: 
\code{\link{dgpoisson}()},
\code{\link{dispersion}()},
\code{\link{fitted,inla-method}},
\code{\link{get_observed}()}
}
\concept{statistics}
