% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa_classes.R, R/rgeneric.R
\docType{class}
\name{rgeneric-class}
\alias{rgeneric-class}
\alias{rgeneric}
\alias{graph.rgeneric}
\alias{prec.rgeneric}
\alias{initial.rgeneric}
\alias{mu.rgeneric}
\alias{prior.rgeneric}
\title{The \code{rgeneric} class organize data needed
to work with Gaussian Markov Random Fields - GMRF,
defined as a R interface for \code{INLA}.}
\usage{
rgeneric(model, debug = FALSE, compile = TRUE, optimize = TRUE, ...)

\method{graph}{rgeneric}(model, ...)

\method{prec}{rgeneric}(model, ...)

\method{initial}{rgeneric}(model)

\method{mu}{rgeneric}(model, theta)

\method{prior}{rgeneric}(model, theta)
}
\arguments{
\item{model}{a \code{rgeneric} model object}

\item{debug}{logical indicating debug state.}

\item{compile}{logical indicating to compile the model.}

\item{optimize}{logical indicating if only the elements
of the precision matrix are returned.}

\item{...}{additional parameter such as 'theta'
If 'theta' is not supplied, initial will be taken.}

\item{theta}{the parameter.}
}
\value{
\code{rgeneric}/ \code{inla.rgeneric} object.
}
\description{
The \code{rgeneric} class organize data needed
to work with Gaussian Markov Random Fields - GMRF,
defined as a R interface for \code{INLA}.
}
\section{Functions}{
\itemize{
\item \code{graph(rgeneric)}: The graph method for 'rgeneric'

\item \code{prec(rgeneric)}: The precision method for an \code{rgeneric} object.

\item \code{initial(rgeneric)}: The initial method for 'rgeneric'

\item \code{mu(rgeneric)}: The mu method for 'rgeneric'

\item \code{prior(rgeneric)}: The prior metho for 'rgeneric'

}}
