% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa_classes.R, R/rgeneric.R
\docType{class}
\name{rgeneric-class}
\alias{rgeneric-class}
\alias{rgeneric}
\alias{rgeneric.default}
\alias{rgeneric.rgeneric}
\alias{rgeneric.inla.rgeneric}
\alias{print.rgeneric}
\alias{summary.rgeneric}
\alias{plot.rgeneric}
\title{Organize data for the latent GMRF R interface for \code{INLA}.}
\usage{
rgeneric(model, debug = FALSE, compile = TRUE, optimize = TRUE, ...)

\method{rgeneric}{default}(model, debug = FALSE, compile = TRUE, optimize = TRUE, ...)

\method{rgeneric}{rgeneric}(model, ...)

\method{rgeneric}{inla.rgeneric}(model, ...)

\method{print}{rgeneric}(x, ...)

\method{summary}{rgeneric}(object, ...)

\method{plot}{rgeneric}(x, y, ...)
}
\arguments{
\item{model}{For the default method, the model defined as a function.
See the 'rgeneric' vignette from the INLA package.}

\item{debug}{logical indicating debug state.}

\item{compile}{logical indicating to compile the model.}

\item{optimize}{logical indicating if only the elements
of the precision matrix are returned.}

\item{...}{not used}

\item{x}{a rgeneric object}

\item{object}{a rgeneric object}

\item{y}{not used}
}
\value{
\code{rgeneric}/ \code{inla.rgeneric} object.
}
\description{
Organize data for the latent GMRF R interface for \code{INLA}.

The rgeneric default method.
}
\section{Functions}{
\itemize{
\item \code{rgeneric(rgeneric)}: Returns the model object unchanged.

\item \code{rgeneric(inla.rgeneric)}: Converts a regular \code{inla.rgeneric} object to \code{rgeneric}.

\item \code{print(rgeneric)}: Print the rgeneric object

\item \code{summary(rgeneric)}: A summary for a rgeneric object

\item \code{plot(rgeneric)}: A plot for a rgeneric object

}}
