% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa_classes.R
\docType{class}
\name{rgeneric-class}
\alias{rgeneric-class}
\title{\code{rgeneric} class to define a \code{\link[INLA:rgeneric]{INLA::rgeneric()}} latent model}
\description{
\code{rgeneric} class to define a \code{\link[INLA:rgeneric]{INLA::rgeneric()}} latent model
}
