% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa_classes.R, R/rgeneric.R
\docType{class}
\name{rgeneric-class}
\alias{rgeneric-class}
\alias{graph.inla.rgeneric}
\alias{prec.inla.rgeneric}
\alias{initial.inla.rgeneric}
\alias{mu.inla.rgeneric}
\alias{prior.inla.rgeneric}
\title{The \code{rgeneric} class for \code{\link[=rgeneric]{rgeneric()}}.}
\usage{
\method{graph}{inla.rgeneric}(model, ...)

\method{prec}{inla.rgeneric}(model, ...)

\method{initial}{inla.rgeneric}(model)

\method{mu}{inla.rgeneric}(model, theta)

\method{prior}{inla.rgeneric}(model, theta)
}
\arguments{
\item{model}{a \code{inla.rgeneric} model object}

\item{...}{additional parameter such as 'theta'
If 'theta' is not supplied, initial will be taken.}

\item{theta}{the parameter.}
}
\description{
The \code{rgeneric} class for \code{\link[=rgeneric]{rgeneric()}}.
}
\section{Functions}{
\itemize{
\item \code{graph(inla.rgeneric)}: The graph method for 'inla.rgeneric'

\item \code{prec(inla.rgeneric)}: The precision method for an \code{inla.rgeneric} object.

\item \code{initial(inla.rgeneric)}: The initial method for 'inla.rgeneric'

\item \code{mu(inla.rgeneric)}: The mu method for 'inla.rgeneric'

\item \code{prior(inla.rgeneric)}: The prior metho for 'inla.rgeneric'

}}
