% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rgeneric.R
\name{rgeneric}
\alias{rgeneric}
\alias{rgeneric.default}
\title{Define rgeneric methods.}
\usage{
rgeneric(model, debug = FALSE, compile = TRUE, optimize = TRUE, ...)

\method{rgeneric}{default}(model, debug = FALSE, compile = TRUE, optimize = TRUE, ...)
}
\arguments{
\item{model}{the model defined as a function.
See the 'rgeneric' vignette from the INLA package.}

\item{debug}{logical indicating debug state.}

\item{compile}{logical indicating to compile the model.}

\item{optimize}{logical indicating if only the elements
of the precision matrix are returned.}

\item{...}{additional arguments to be used internally
for the model, for example, additional data.}
}
\value{
a \code{inla.rgeneric} object.
}
\description{
Define rgeneric methods.

The rgeneric default method.
}
