% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zeroinflated_poisson.R
\name{rzipois}
\alias{rzipois}
\title{Zero-inflated Poisson}
\usage{
rzipois(n, lambda, prob)
}
\arguments{
\item{n}{number of samples}

\item{lambda}{mean of the Poisson distribution}

\item{prob}{the mass of extra zero's}
}
\description{
Generate random numbers from a zero-inflated Poisson distribution
}
