% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Matrix.R
\name{Sparse}
\alias{Sparse}
\title{To store in i,j,x sparse matrix format}
\usage{
Sparse(A, unique = TRUE, na.rm = FALSE, zeros.rm = FALSE)
}
\arguments{
\item{A}{matrix or Matrix}

\item{unique}{logical (default is TRUE) to
ensure that the internal representation
is unique and there are no duplicated entries.
(Do not change this unless you know what you are doing.)}

\item{na.rm}{logical (default is FALSE) indicating
if it is to replace ‘NA’'s in the matrix with zeros.}

\item{zeros.rm}{logical (default is FALSE)
indicating if it is to remove zeros in the
matrix. Applied after \code{na.rm}.}
}
\description{
To store in i,j,x sparse matrix format
}
\note{
This is based in INLA::inla.as.sparse(),
but allow all combinations of 'na.rm' and 'zeros.rm'.
}
